Removed unnecessary DataObject registration
