--- java/languages.antlr/src/org/netbeans/modules/languages/antlr/AntlrDataObject.java
@@ -1,139 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.netbeans.modules.languages.antlr;
-
-import org.netbeans.core.spi.multiview.MultiViewElement;
-import org.netbeans.core.spi.multiview.text.MultiViewEditorElement;
-import org.openide.awt.ActionID;
-import org.openide.awt.ActionReference;
-import org.openide.awt.ActionReferences;
-import org.openide.filesystems.FileObject;
-import org.openide.filesystems.MIMEResolver;
-import org.openide.loaders.DataNode;
-import org.openide.loaders.DataObject;
-import org.openide.loaders.DataObjectExistsException;
-import org.openide.loaders.MultiDataObject;
-import org.openide.loaders.MultiFileLoader;
-import org.openide.nodes.Children;
-import org.openide.nodes.Node;
-import org.openide.util.Lookup;
-import org.openide.util.NbBundle;
-import org.openide.windows.TopComponent;
-
-/**
- *
- * @author lkishalmi
- */
-@NbBundle.Messages(
-        "ANTLRResolver=ANTLR4 Grammar"
-)
-@MIMEResolver.ExtensionRegistration(displayName = "#ANTLRResolver",
-    extension = "g4",
-    mimeType = AntlrTokenId.MIME_TYPE,
-    position = 285
-)
-
-@ActionReferences({
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "System", id = "org.openide.actions.OpenAction"),
-            position = 100,
-            separatorAfter = 300
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "Edit", id = "org.openide.actions.CutAction"),
-            position = 400
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "Edit", id = "org.openide.actions.CopyAction"),
-            position = 500,
-            separatorAfter = 600
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "Edit", id = "org.openide.actions.DeleteAction"),
-            position = 700
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "System", id = "org.openide.actions.RenameAction"),
-            position = 800,
-            separatorAfter = 900
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "System", id = "org.openide.actions.SaveAsTemplateAction"),
-            position = 1000,
-            separatorAfter = 1100
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "System", id = "org.openide.actions.FileSystemAction"),
-            position = 1200,
-            separatorAfter = 1300
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "System", id = "org.openide.actions.ToolsAction"),
-            position = 1400
-    ),
-    @ActionReference(
-            path = "Loaders/text/x-antlr4/Actions",
-            id = @ActionID(category = "System", id = "org.openide.actions.PropertiesAction"),
-            position = 1500
-    )
-})
-
-@DataObject.Registration(
-        mimeType = AntlrTokenId.MIME_TYPE,
-        displayName = "#ANTLRResolver",
-        position = 304
-)
-public class AntlrDataObject extends MultiDataObject {
-
-    public AntlrDataObject(FileObject fo, MultiFileLoader loader) throws DataObjectExistsException {
-        super(fo, loader);
-        registerEditor(AntlrTokenId.MIME_TYPE, true);
-    }
-
-    @Override
-    protected Node createNodeDelegate() {
-        return new DataNode(this, Children.LEAF, getLookup());
-    }
-
-    @Override
-    protected int associateLookup() {
-        return 1;
-    }
-
-    @NbBundle.Messages("Source=&Source")
-    @MultiViewElement.Registration(
-            displayName = "#Source",
-            persistenceType = TopComponent.PERSISTENCE_ONLY_OPENED,
-            mimeType = AntlrTokenId.MIME_TYPE,
-            preferredID = "neon.source",
-            position = 1
-    )
-    public static MultiViewEditorElement createMultiViewEditorElement(Lookup context) {
-        return new MultiViewEditorElement(context);
-    }
-
-}
